+++ package.json
@@ -15,6 +15,10 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "@types/jest": "^26.0.20",
+    "@types/node": "^14.14.31",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.1",
     "concurrently": "^6.0.0",
     "cors": "^2.8.5",
     "custom-electron-titlebar": "3.2.5",
@@ -32,6 +36,7 @@
     "redux": "^4.0.5",
     "socket.io": "^3.1.1",
     "socket.io-client": "^3.1.1",
+    "typescript": "^4.2.3",
     "wait-on": "^5.2.1",
     "web-vitals": "^1.0.1",
     "websocket": "^1.0.33"
@@ -64,6 +69,8 @@
     ]
   },
   "devDependencies": {
+    "@types/react-redux": "^7.1.16",
+    "@types/react-router-dom": "^5.1.7",
     "electron-devtools-installer": "^3.1.1",
     "react-app-rewired": "^2.1.8",
     "redux-socket.io-middleware": "^1.0.4"
