+++ src/application/actions/provider.js
@@ -5,6 +5,7 @@ export const CONNECT_PROVIDER_FAILURE       = '[provider] connect failure';
 export const CALL_GAMEPOWER_EXTRINSIC       = 'provider://gamepower';
 export const CALL_NFT_EXTRINSIC             = 'provider://nft';
 export const CALL_SYSTEM_EXTRINSIC          = 'provider://system';
+export const CALL_GAMEPOWER_STORAGE         = 'provider://gamepower/storage';
 
 export const connectToProvider = {
     type: CONNECT_PROVIDER,
@@ -34,3 +35,8 @@ export const callSystemExtrinsic = data => ({
     type: CALL_SYSTEM_EXTRINSIC,
     payload: data,
 });
+
+export const callGamepowerStorage = data => ({
+    type: CALL_GAMEPOWER_STORAGE,
+    payload: data,
+})
