+++ src/application/actions/ui.js
@@ -1,10 +1,15 @@
-export const APP_INIT = '[ui] app init';
-export const SET_LOADING_ON = '[ui] set loading on';
-export const SET_LOADING_OFF = '[ui] set loading off';
-export const SET_LOADING_MESSAGE = '[ui] set loading message';
+export const APP_INIT                       = '[ui] app init';
+export const SET_LOADING_ON                 = '[ui] set loading on';
+export const SET_LOADING_OFF                = '[ui] set loading off';
+export const SET_LOADING_MESSAGE            = '[ui] set loading message';
+export const CREATE_MENUBAR                 = '[ui] create menubar';
 
 export const appInit = {
     type: APP_INIT
+};
+
+export const createMenuBar = {
+    type: CREATE_MENUBAR
 };
 
 export const setLoading = isLoading => ({
