+++ src/application/middleware/provider.js
@@ -8,7 +8,7 @@ const connectProviderFlow = ({api}) => ({dispatch}) => next => async (action) =>
         try {
             dispatch(uiActions.setLoadingMessage("Connecting to blockchain"));
             dispatch(uiActions.setLoading(true));
-            const connection = await api.provider.connect("wss://gamepower.io");
+            const connection = await api.provider.connect(/*"wss://gamepower.io"*/);
             dispatch(providerActions.connectProviderSuccess(connection));
             dispatch(uiActions.setLoading(false));
         } catch (error) {
@@ -24,10 +24,23 @@ const callGamerpowerExtrinsicFlow = ({api, log}) => ({getState, dispatch}) => ne
         try {
             const account = getState().keyring.activeAccount;
             const connection = getState().provider.connection;
-            console.log(action.payload);
             api.provider.gamepower(dispatch, connection, account, action.payload);
         } catch (error) {
             //dispatch(connectProviderFailure(error));
+        }
+    }
+}
+
+const callGamerpowerStorageFlow = ({api, log}) => ({getState, dispatch}) => next => async (action) => {
+    next(action);
+
+    if (action.type === providerActions.CALL_GAMEPOWER_STORAGE) {
+        try {
+            const account = getState().keyring.activeAccount;
+            const connection = getState().provider.connection;
+            api.provider.queryGamepower(dispatch, connection, account);
+        } catch (error) {
+            log(error);
         }
     }
 }
@@ -51,4 +64,5 @@ export default [
     connectProviderFlow,
     callGamerpowerExtrinsicFlow,
     callSystemExtrinsicFlow,
+    callGamerpowerStorageFlow,
 ]
