+++ src/application/middleware/ui.js
@@ -8,13 +8,23 @@ const appInitFlow = ({ api, log }) => ({ dispatch }) => next => action => {
     if (action.type === uiActions.APP_INIT) {
         log('app init');
         // Create menubar
-        api.menubar.createMenuBar();
+        dispatch(uiActions.createMenuBar);
 
         // Connect to the blockchain
         dispatch(providerActions.connectToProvider);
     }
 }
 
+const createMenubarFlow = ({ api, log }) => ({ dispatch }) => next => action => {
+    next(action);
+
+    if (action.type === uiActions.CREATE_MENUBAR) {
+        // Create menubar
+        api.menubar.createMenuBar();
+    }
+}
+
 export default [
-    appInitFlow
+    appInitFlow,
+    createMenubarFlow
 ]
