+++ src/infrastructure/interfaces/IAccount.tsx
@@ -1,0 +1,14 @@
+export interface IAccount {
+    address?: string;
+    meta?: IAccountMeta;
+    publicKey?: Uint8Array;
+    addressRaw?: Uint8Array;
+    isLocked?: Boolean;
+    type?: string;
+    accountInfo?: any;
+}
+
+export interface IAccountMeta {
+    name: string;
+    whenCreated: number;
+}
