+++ src/infrastructure/interfaces/IInventoryItem.tsx
@@ -1,0 +1,6 @@
+export interface IInventoryItem {
+    title?: string;
+    creator?: string;
+    image?: string;
+    price?: number;
+}
