+++ src/infrastructure/interfaces/IState.tsx
@@ -1,0 +1,5 @@
+import { IKeyRingState } from "./IKeyRingState";
+
+export interface IState {
+    keyring: IKeyRingState;
+}
