--- src/views/signup/Signup.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Signup() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
