+++ src/views/signup/Signup.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Signup() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
